assignment 2/libs/exceptions/global_exception_handler.py
from rest_framework.views import exception_handler

# TODO: improve response error
def handler(exc, context):
    response = exception_handler(exc, context)

    # Now add the HTTP status code to the response.
    if response is not None:
        response.data['status_code'] = response.status_code

    return response